+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [0.4.5] 2023-08-23
 
+- fix metadata in README.md
+- write out nextflow_*.json files
+- add summarize wildcards for QuiltProduct
 - reduce logging (especially stack traces and filesystem calls)
 - convert non-fatal errors to warnings